requirements.txt
fastapi
uvicorn
requests
supabase
google-api-python-client
google-auth-oauthlib
pydantic
python-dotenv
